Change "license" to "licence" but keep correct mock server and api. !Watch out, the names of the config properties for licence-connect-authenticator are being changed"
