+++ docker-compose.yaml
@@ -27,8 +27,8 @@ services:
       KC_SPI_ADMIN_REALM_RESTAPI_EXTENSION_HMAC_VERIFIER_REALM: fwu
       KC_SPI_ADMIN_REALM_RESTAPI_EXTENSION_HMAC_MANAGEMENT_REALM: management
       KC_SPI_EVENTS_LISTENER_REMOVE_USER_ON_LOGOUT_FWU: "IDP"
-      KC_SPI_REMOVE_USER_REST_CLIENT_DEFAULT_LICENSE_CONNECT_BASE_URL: http://mockserver:1080/v1/licences/release
-      KC_SPI_REMOVE_USER_REST_CLIENT_DEFAULT_LICENSE_CONNECT_API_KEY: sample-api-key
+      KC_SPI_REMOVE_USER_REST_CLIENT_DEFAULT_LICENCE_CONNECT_BASE_URL: http://mockserver:1080/v1/licences/release
+      KC_SPI_REMOVE_USER_REST_CLIENT_DEFAULT_LICENCE_CONNECT_API_KEY: sample-api-key
       KC_SPI_ADMIN_REALM_RESTAPI_EXTENSION_VIDIS_CUSTOM_FWU: "ALL"
       KC_SPI_ADMIN_REALM_RESTAPI_EXTENSION_VIDIS_CUSTOM_DELETIONTOLERANCE: 5
       KC_FEATURES: admin-api
