+++ licence-connect-authenticator/src/main/resources/META-INF/services/org.keycloak.authentication.AuthenticatorFactory
@@ -1,0 +1,1 @@
+de.intension.authentication.authenticators.licence.LicenceConnectAuthenticatorFactory
