+++ licence-connect-authenticator/src/test/resources/idp-realm-licence.json
