--- license-connect-authenticator/src/main/resources/META-INF/services/org.keycloak.authentication.AuthenticatorFactory
@@ -1,1 +1,0 @@
-de.intension.authentication.authenticators.license.LicenseConnectAuthenticatorFactory
