--- remove-user-on-logout/src/main/java/de/intension/rest/model/RemoveLicenseRequest.java
@@ -1,22 +1,0 @@
-package de.intension.rest.model;
-
-public class RemoveLicenseRequest
-{
-
-    private String studentId;
-
-    public RemoveLicenseRequest(String studentId)
-    {
-        this.studentId = studentId;
-    }
-
-    public String getStudentId()
-    {
-        return studentId;
-    }
-
-    public void setStudentId(String studentId)
-    {
-        this.studentId = studentId;
-    }
-}
