+++ remove-user-on-logout/src/main/java/de/intension/spi/RestClientProvider.java
@@ -2,12 +2,12 @@ package de.intension.spi;
 
 import org.keycloak.provider.Provider;
 
-import de.intension.rest.LicenseConnectRestClient;
+import de.intension.rest.LicenceConnectRestClient;
 
 public interface RestClientProvider
     extends Provider
 {
 
-    LicenseConnectRestClient restClient();
+    LicenceConnectRestClient restClient();
 
 }
